* IIOP/test/svr.cpp: Global function ::OA_listen() no longer exists; its functionality is now fully contained within ::main(). If USE_ACE_EVENT_HANDLING is defined, a Reactor-based event loop is used.  The original intent was to allow conditional compilation to select btw. the original method and a Reactor-based method. However, weaving that into the library proved far more difficult than anticipated, so more than likely not defining USE_ACE_EVENT_HANDLING will cause massive grief.
* IIOP/lib/toa.cpp: Changed call to TCP_OA::init() to reflect new
ACE_INET_Addr argument type.

* IIOP/lib/tcpoa.cpp (TCP_OA): All initialization methods were
changed, specifically the CTOR and TCP_OA::init, to reflect the
introduction of ACE_INET_Addr.  Two side-effects of using
ACE_INET_Addr are that (1) a server can specify the address on
which it wants to listen and (2) best that I can tell, the server
MUST specify the address on which it wants to listen because
otherwise it won't be able to publish a rational IOR.  The event
loop is now changed to simply loop on Reactor::handle_events().

* IIOP/lib/tcpoa.hh (TCP_OA): Removed vestiges of this component's
use of the original connection management scheme.  Where
appropriate, hostnames and ports were replaces by ACE_INET_Addr,
endpoints by ACE_HANDLEs, etc.  One particularly nasty thing done
was to declare ROA_Handler as a friend so that handle_message()
can be called from ROA_Handler::handle_input(), which to me
exposes a hole in the original architecture wherein input is
"pulled" rather than waited-for.  We might need to re-think how
this is handled within TAO.

* IIOP/lib/roa.{hh,i,cpp}: These files contain the required
components to support the new server-side ACE-based
connection/event substrate.  The client side remains, as always,
using the connection mgmt scheme used by the original Sun IIOP
code.

* IIOP/lib/giop.cpp: Added explicit template instantiation for
when this is needed.

* IIOP/lib/corbacom.hh: Now protects itself from multiple
inclusion.

* IIOP/lib/Makefile: Added roa.* where appropriate.

* IIOP/test/test1_{clnt,svr}.cpp: THESE HAVE NOT BEEN CONVERTED TO
USE ACE EVENT HANDLING!!  This round of changes explicitly
instantiates templates where required.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk/TAO@1377 ae88bc3d-4319-0410-8dbf-d08b4c9d3795