* tao/iioporb.cpp: Added call to set object address in Profile.
* tao/iiopobj.cpp: Corrected code in the copy CTOR where the host
was being copied over the old host.  This could prove bad if the
amount of space allocated previously was too small for the
hostname we're copying in.  Ideally we'd use something smarter
than this, but hopefully (if you read the comment below) this will
go away soon.

* tao/giop.cpp: Use the IIOP::Profile object address when making a
connection.

* tao/decode.cpp: Added code to set the object address in the
Profile when an object ref is decoded.

* tao/debug.h: Removed old questions in comments.

* tao/connect.h: Added some documentation.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk/TAO@5829 ae88bc3d-4319-0410-8dbf-d08b4c9d3795