* tests/Cubit/TAO/IDL_Cubit/clnt.cpp: Changed the assignment of default port number to fetch the value from the environment similar to what is done in orb_core.cpp.
* tao/orb_core.cpp: Added the ability to specify the default
server port via an environment variable using the ACE_Env_Value
template.  The name of the environment variable is the same as the
preprocessor #define: TAO_DEFAULT_SERVER_PORT.

* tao/Makefile: Added "fakesvcconf" variable check so that if
somebody wants to build with faked svc config entries on a
platform, they can simply do "make fakesvcconf=1".  The CHORUS and
VXWORKS sections now override the value of this variable to be 1.

* Added the ability to understand and respond to the GIOP
LocateRequest message.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk/TAO@7233 ae88bc3d-4319-0410-8dbf-d08b4c9d3795