* tao/strategy_T.h: Added this file to contain template strategies such as TAO_Reactive_Strategy<>.
* tao/sequence_T.cpp (operator =): Corrected a problem where T**
was being assigned to T*.

* tao/orb_core.cpp (init): Moved setting of stuff in orb_params()
to be post-service initialization.  It needs to be this way
because the ORB Parameters are now regulated by the resource
factory settings, and that gets created during service
initialization.

* tao/giop.cpp: Changed the TAO_GIOP_LocateRequestHeader::init to
be more efficient and not generate warnings on Irix.

* tao/giop.h: Change ORB-internal sequences (such as the
ServiceContextList) to use the new TAO_Unbounded_Sequence<>
template rather than the deprecated CORBA_SEQUENCE<> hack.


git-svn-id: svn://svn.dre.vanderbilt.edu/DOC/Middleware/trunk/TAO@7559 ae88bc3d-4319-0410-8dbf-d08b4c9d3795